fix: fix next image | missing session instance on complete profile
